Spi fixed w.r.t CS line assertion for Jobs
CS line was asserted for every channel transaction is now corrected to
assertion done for every job. This allows us to read registers of
ENC28J60 with Motorola SPI format itself. Also found that NS Microwire
format has 1 extra clock making it to be not usable for ENC28J60 driver.